Animation on item add
* That's (part of) why we switch to RecyclerView.
But... some bug: when notifyItemInserted, the list is not scrolled to top (it remembers where it is) and we can't see the animation as a result.
However, we are using S2R so when we are refreshing we are certainly at the top.
Some workaround is used. No problem so far. Maybe it's by design and this
is the correct way?